requirements.txt
fastapi
uvicorn[standard]
python-dotenv
neo4j
langchain
langchain-community
langchain-openai
tiktoken
duckdb
pandas
unstructured
unstructured-inference
layoutparser
reportlab
pdfplumber
prometheus-client